Update tests for SDK based builds.
