+++ test/dotnet-shift.Tests/DeployHandlerTests.SuccessfulDeploymentWithFollow.verified.txt
@@ -3,9 +3,12 @@
 Retrieving existing resources...
 
 Updating build resources...
-Uploading sources from directory '{TempPath}empty-project'...
-The build is running:
-Application built successfully.
+
+Building .NET project '{TempPath}empty-project/project.csproj'...
+Running dotnet publish {TempPath}empty-project/project.csproj.
+
+Publishing application image to image registry...
+Succesfully pushed image 'image-registry.openshift-image-registry.svc:5000/mynamespace/mycomponent@sha256:deadbeef'
 
 Updating deployment resources...
 The image is being deployed.
